pyproject.toml
[tool.poetry]
name = "octoxlabscase"
version = "0.1.0"
description = ""
authors = ["yahya kemal <yahyakemalkus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.0.3"
ruff = "^0.3.2"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
uuid = "^1.30"
pyjwt = "^2.8.0"
urllib3 = "^2.2.1"
elasticsearch = "^8.12.1"
pydantic = "1.10.4"
pytest = "^8.1.1"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "octoxlabscase.settings"
